Removing a few warnings and simplifying bestScore omputation in Nominal.cpp

+++ particle_filter/PFWatcher.hpp
@@ -77,7 +77,7 @@ private:
 
     void removeExcess()
         {
-            while (entries.size() > windowSize) {
+            while ((int)entries.size() > windowSize) {
                 entries.pop_front();
             }
         }
